requirements.txt
colorama==0.4.4
DateTime==4.3
pause==0.2
python-dotenv==0.14.0
pytz==2020.1
selenium==3.141.0
termcolor==1.1.0
urllib3==1.25.11
zope.interface==5.1.2
